+++ src/Model/BaseModel.php
@@ -1,16 +1,22 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: PBX_g33k
- * Date: 31-May-16
- * Time: 01:04
- */
+/*******************************************************************************
+ * This file is part of the Pbxg33k\MusicInfo package.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * (c) 2017 Oguzhan uysal. All rights reserved
+ ******************************************************************************/
 
 namespace Pbxg33k\MusicInfo\Model;
 
 
-class BaseModel
+use Pbxg33k\Traits\PropertyTrait;
+
+abstract class BaseModel
 {
+    use PropertyTrait;
+
     /**
      * Data source identifier
      *
@@ -54,7 +60,7 @@ class BaseModel
     }
 
     /**
-     * @param array $rawData
+     * @param mixed $rawData
      *
      * @return BaseModel
      */