+++ src/MusicInfo.php
@@ -1,4 +1,13 @@
 <?php
+/*******************************************************************************
+ * This file is part of the Pbxg33k\MusicInfo package.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * (c) 2017 Oguzhan uysal. All rights reserved
+ ******************************************************************************/
+
 namespace Pbxg33k\MusicInfo;
 
 use Doctrine\Common\Collections\ArrayCollection;
@@ -244,6 +253,7 @@ class MusicInfo
             if (!method_exists($service, $methodName)) {
                 throw new \Exception(sprintf('Method (%s) not found in %s', $methodName, get_class($service)));
             }
+
             $results->set($serviceKey, $service->{$methodName}()->getByName($argument));
         }